SelectFieldFromApi should handle array response
SelectFieldFromApi should handle the case where the response body of the
request is already an array. As it is arraySelector is required but there
is no means of pointing to the root with a selector AFAICS. This change
defaults the data to the response body if the arraySelector is not set.