ability to filter users/groups in okta provider
In order to do this, I needed to create a new provider that does both the groups
and users together in the same provider. This is because if you filter a group, you
will likely want to filter its members. Likewise, if you filter all users from a group
you likely want to exclude the group.