add lineWidth option to yaml.dump for JSONata and serialization actions
The JSONata action is very useful for modifying an existing YAML file.
However, it fails to match the style of the destination YAML. If you
have a file that looks like this:

```yaml
someKey: moreThan80Chars
```

and run a transformation on it (`$ ~> | $ | { "newKey": "newVal" } |`),
the end result is something like:

```yaml
someKey: >-
  moreThan80Chars
newKey: newval
```

For a file with many such lines (in our case, AWS ARNs), this creates a
massive diff against the file, when only a small transformation was
performed.

This change allows for setting the `lineWidth` property of
`js-yaml.dump`[0], which when set to `-1`, disables auto-wrapping. This
would allow us to use JSONata transformations without enforcing the
default `js-yaml` style on the destination YAML.

[0]: https://www.npmjs.com/package/js-yaml

Signed-off-by: Riley Martine <rmartine@integralads.com>