Add option to merge array values  (#1201)
When merging objects array values are overwritten by default. This change allows the user to specify that they'd like the arrays merged.
With this option enabled the values in the target array will be concatenated with the merge value. This change does not attempt to merge items within the arrays.