Allow unsetting SelectFieldFromApi (#1309)
Previously when the user selected the placeholder value
the value was set to an empty array. This lead to and error
validating the template parameters due to inconsistent types
array vs string. This change unsets the value if the value returned
is an empty array. We don't currently support multiple select so an
array isn't a valid value currently.