Fix Jira requests for new authentication system
In the latest releases the authentication system has been updated
and improved a lot. However, this is now breaking the Jira plugin.

To fix it, I've changed the old implementation using `fetch` and `identityApi` to
inject the token in the headers with directly the `fetchApi`, which does it
automatically for you. Once this has been done, the error 403 obtained in the "search"
request is gone.

Also updated the tests to inject the new fetchApi instead of the identityApi and simplified
them a little, using some new mechanisms for testing frontend